Tidy Gapminder indicator data sheets stored in a folder using tidy_indice.

Tidy Gapminder indicator data sheets stored in a folder using tidy_indice.

tidy_bunch(dirpath = ".", merge = FALSE, ...)

Arguments

dirpath

Path to folder containing indices data sheets.

merge

A boolean. If true, combines all indices data sheets in one. Default: FALSE.

...

Ellipsis.

Value

List of indices data frames tidied (if merge = FALSE) or a single data frame of all indices tidied (if merge = TRUE).

Examples

folder_path <- system.file("extdata", package = "tidygapminder")

tidy_bunch(folder_path)
#> [[1]]
#> # A tibble: 11,076 × 3
#>    country      year `Agricultural land (% of land area)`
#>    <chr>       <dbl>                                <dbl>
#>  1 Afghanistan  1960                                 NA  
#>  2 Afghanistan  1961                                 57.8
#>  3 Afghanistan  1962                                 57.9
#>  4 Afghanistan  1963                                 58.0
#>  5 Afghanistan  1964                                 58.1
#>  6 Afghanistan  1965                                 58.1
#>  7 Afghanistan  1966                                 58.1
#>  8 Afghanistan  1967                                 58.2
#>  9 Afghanistan  1968                                 58.2
#> 10 Afghanistan  1969                                 58.3
#> # ℹ 11,066 more rows
#> 
#> [[2]]
#> # A tibble: 40,953 × 3
#>    country      year life_expectancy_years
#>    <chr>       <dbl>                 <dbl>
#>  1 Afghanistan  1800                  28.2
#>  2 Afghanistan  1801                  28.2
#>  3 Afghanistan  1802                  28.2
#>  4 Afghanistan  1803                  28.2
#>  5 Afghanistan  1804                  28.2
#>  6 Afghanistan  1805                  28.2
#>  7 Afghanistan  1806                  28.1
#>  8 Afghanistan  1807                  28.1
#>  9 Afghanistan  1808                  28.1
#> 10 Afghanistan  1809                  28.1
#> # ℹ 40,943 more rows
#>